refactor: migrate state handler implementations to core/handlers and update imports
