+++ core/handlers/lambda_handler.py
