+++ core/parser_machine.py
@@ -4,9 +4,9 @@ from typing import Any, Callable
 import logging
 import yaml
 from logging_config import setup_logging
-from core.blocks.choice_handler import Choice
-from core.blocks.lambda_handler import Lambda
-from core.blocks.parallel_handler import Parallel
+from core.handlers.choice_handler import Choice
+from core.handlers.lambda_handler import Lambda
+from core.handlers.parallel_handler import Parallel
 from core.state_machine import StateMachine
